slug: RGBA32Uint textures for WebGPU (three has no 16-bit integer path)
WebGPU rejected the slug curve/glyph-map textures two ways:
  1. texture.internalFormat='RGBA16UI' — a WebGL2 hint the WebGPU backend
     passes through verbatim as a GPUTextureFormat (invalid; formats are
     lowercase enums). Dropped it; let three derive the format.
  2. three's WebGPU getFormat() maps RGBAIntegerFormat only to 32-bit
     (Int->RGBA32Sint, UnsignedInt->RGBA32Uint); there's no UnsignedShort
     path. So RGBA16UI never created.

Fix: widen to RGBA32Uint — Uint32Array data + UnsignedIntType. Packing
(0..65535) and the TSL shader (textureLoad + /65535) are unchanged, so this
isolates the texture-format fix from the shader port. Deleted the now-dead
SLUG_TEXTURE_FORMAT doc-const and refreshed slug-constants header.

+++ src/shaping/slugData.js
@@ -66,12 +66,14 @@ export function buildSlugBuffers(shaper, glyphIds) {
     const totalCurveTexels = totalCurves * CURVE_TEXELS_PER_CURVE;
     const curveTexHeight = Math.max(1, Math.ceil(totalCurveTexels / TEXTURE_WIDTH));
     const curveTexSize = TEXTURE_WIDTH * curveTexHeight * 4; // 4 channels (RGBA)
-    const curveData = new Uint16Array(curveTexSize);
+    // Uint32 (RGBA32Uint): three's WebGPU backend has no 16-bit integer texture
+    // path — RGBAIntegerFormat only maps Int/UnsignedInt. Values stay 0..65535.
+    const curveData = new Uint32Array(curveTexSize);
 
     // GlyphMap texture: 1 texel per glyph slot
     const glyphMapTexHeight = Math.max(1, Math.ceil(glyphMapEntries / TEXTURE_WIDTH));
     const glyphMapTexSize = TEXTURE_WIDTH * glyphMapTexHeight * 4;
-    const glyphMapData = new Uint16Array(glyphMapTexSize);
+    const glyphMapData = new Uint32Array(glyphMapTexSize);
 
     // Phase 3: Pack data into texture arrays
     let curveTexelOffset = 0; // current texel write position in curveTexture
